+++ bindings/nostr-ffi/src/event/mod.rs
@@ -3,7 +3,6 @@
 // Distributed under the MIT software license
 
 use std::ops::Deref;
-use std::sync::Arc;
 
 use nostr::JsonUtil;
 use uniffi::Object;
@@ -18,10 +17,9 @@ pub mod unsigned;
 pub use self::builder::EventBuilder;
 pub use self::id::EventId;
 pub use self::kind::{Kind, KindEnum};
-pub use self::tag::{Tag, TagKind, TagStandard};
+pub use self::tag::{Tag, TagKind, TagStandard, Tags};
 pub use self::unsigned::UnsignedEvent;
 use crate::error::Result;
-use crate::nips::nip01::Coordinate;
 use crate::{PublicKey, Timestamp};
 
 #[derive(Debug, PartialEq, Eq, Hash, Object)]
@@ -67,29 +65,8 @@ impl Event {
         self.inner.kind.into()
     }
 
-    pub fn tags(&self) -> Vec<Arc<Tag>> {
-        self.inner
-            .tags
-            .iter()
-            .cloned()
-            .map(|t| Arc::new(t.into()))
-            .collect()
-    }
-
-    /// Get content of **first** tag that match `TagKind`.
-    pub fn get_tag_content(&self, kind: TagKind) -> Option<String> {
-        self.inner
-            .get_tag_content(kind.into())
-            .map(|c| c.to_string())
-    }
-
-    /// Get content of all tags that match `TagKind`.
-    pub fn get_tags_content(&self, kind: TagKind) -> Vec<String> {
-        self.inner
-            .get_tags_content(kind.into())
-            .into_iter()
-            .map(|c| c.to_string())
-            .collect()
+    pub fn tags(&self) -> Tags {
+        self.inner.tags.clone().into()
     }
 
     #[inline]
@@ -121,64 +98,13 @@ impl Event {
         self.inner.verify_signature()
     }
 
-    /// Get `Timestamp` expiration if set
-    pub fn expiration(&self) -> Option<Arc<Timestamp>> {
-        self.inner.expiration().map(|t| Arc::new((*t).into()))
-    }
-
     /// Returns `true` if the event has an expiration tag that is expired.
-    /// If an event has no `Expiration` tag, then it will return `false`.
+    /// If an event has no expiration tag, then it will return `false`.
     ///
     /// <https://github.com/nostr-protocol/nips/blob/master/40.md>
     #[inline]
     pub fn is_expired(&self) -> bool {
         self.inner.is_expired()
-    }
-
-    /// Extract identifier (`d` tag), if exists.
-    #[inline]
-    pub fn identifier(&self) -> Option<String> {
-        self.inner.identifier().map(|i| i.to_string())
-    }
-
-    /// Extract public keys from tags (`p` tag)
-    ///
-    /// **This method extract ONLY supported standard variants**
-    pub fn public_keys(&self) -> Vec<Arc<PublicKey>> {
-        self.inner
-            .public_keys()
-            .copied()
-            .map(|p| Arc::new(p.into()))
-            .collect()
-    }
-
-    /// Extract event IDs from tags (`e` tag)
-    ///
-    /// **This method extract ONLY supported standard variants**
-    pub fn event_ids(&self) -> Vec<Arc<EventId>> {
-        self.inner
-            .event_ids()
-            .copied()
-            .map(|p| Arc::new(p.into()))
-            .collect()
-    }
-
-    /// Extract coordinates from tags (`a` tag)
-    ///
-    /// **This method extract ONLY supported standard variants**
-    pub fn coordinates(&self) -> Vec<Arc<Coordinate>> {
-        self.inner
-            .coordinates()
-            .cloned()
-            .map(|p| Arc::new(p.into()))
-            .collect()
-    }
-
-    /// Extract hashtags from tags (`t` tag)
-    ///
-    /// **This method extract ONLY supported standard variants**
-    pub fn hashtags(&self) -> Vec<String> {
-        self.inner.hashtags().map(|t| t.to_owned()).collect()
     }
 
     /// Check if it's a protected event