+++ crates/nostr-database/src/lib.rs
@@ -206,7 +206,7 @@ pub trait NostrDatabaseExt: NostrDatabase {
             .limit(1);
         let events: Vec<Event> = self.query(vec![filter]).await?;
         match events.first() {
-            Some(event) => Ok(event.public_keys().copied().collect()),
+            Some(event) => Ok(event.tags.public_keys().copied().collect()),
             None => Ok(Vec::new()),
         }
     }
@@ -223,7 +223,7 @@ pub trait NostrDatabaseExt: NostrDatabase {
             Some(event) => {
                 // Get contacts metadata
                 let filter = Filter::new()
-                    .authors(event.public_keys().copied())
+                    .authors(event.tags.public_keys().copied())
                     .kind(Kind::Metadata);
                 let mut contacts: HashSet<Profile> = self
                     .query(vec![filter])
@@ -237,7 +237,7 @@ pub trait NostrDatabaseExt: NostrDatabase {
                     .collect();
 
                 // Extend with missing public keys
-                contacts.extend(event.public_keys().copied().map(Profile::from));
+                contacts.extend(event.tags.public_keys().copied().map(Profile::from));
 
                 Ok(contacts.into_iter().collect())
             }