+++ crates/nwc/src/lib.rs
@@ -117,7 +117,7 @@ impl NWC {
             while let Ok(notification) = notifications.recv().await {
                 if let RelayNotification::Event { event, .. } = notification {
                     if event.kind == Kind::WalletConnectResponse
-                        && event.event_ids().next().copied() == Some(event_id)
+                        && event.tags.event_ids().next().copied() == Some(event_id)
                     {
                         return Ok(Response::from_event(&self.uri, &event)?);
                     }